feat(home): refactoring of the home component
